Fix invalid usage of aria-labelledby
`aria-labelledby` expects an ID that references another element, the
contents of which are to be used as the label. Here, we use "robot",
"selfish", and "guest", but these are not references to named elements,
but text values.

Instead, `aria-label` should be used. However, `aria-label` does not
reveal its contents to normal users (only screen readers) and there is
improved discoverability for new KGS users as to what `~` means, for
instance, so using `title` (which screen readers understand) providers
both an accessible label and some discoverable name to all users.